% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCT.R
\name{CCT}
\alias{CCT}
\title{To conduct ACAT method}
\usage{
CCT(pvals, weights = NULL)
}
\arguments{
\item{pvals}{the combined p-values}

\item{weights}{the weights}
}
\value{
pval
}
\description{
To conduct ACAT method
}
