% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBJ.R
\name{eBJ}
\alias{eBJ}
\title{To conduct eBJ test}
\usage{
eBJ(Z_score, Sigma)
}
\arguments{
\item{Z_score}{GWAS summary statistics}

\item{Sigma}{the correlation matrix}
}
\value{
pval.eBJ
}
\description{
To conduct eBJ test
}
