% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eHC.R
\name{eHC}
\alias{eHC}
\title{To conduct eHC test}
\usage{
eHC(Z_score, Sigma)
}
\arguments{
\item{Z_score}{GWAS summary statistics}

\item{Sigma}{the correlation matrix}
}
\value{
pval.eHC
}
\description{
To conduct eHC test
}
