% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Ana.R
\name{Eigen_ana}
\alias{Eigen_ana}
\title{To perform multiple trait analysis}
\usage{
Eigen_ana(Z_score, Sigma)
}
\arguments{
\item{Z_score}{GWAS summary statistics}

\item{Sigma}{the correlation matrix}
}
\value{
the p-values of multi-trait tests
}
\description{
To perform multiple trait analysis
}
