% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHC.R
\name{iHC}
\alias{iHC}
\title{To conduct iHC}
\usage{
iHC(Z_score, Sigma)
}
\arguments{
\item{Z_score}{GWAS summary statistics}

\item{Sigma}{the correlation matrix}
}
\value{
pval.iHC
}
\description{
To conduct iHC
}
