% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCFisher.R
\name{PCFisher}
\alias{PCFisher}
\title{To conduct PCFisher test}
\usage{
PCFisher(Z_score, Sigma)
}
\arguments{
\item{Z_score}{GWAS summary statistics}

\item{Sigma}{the correlation matrix}
}
\value{
PC.Fisher.p
}
\description{
To conduct PCFisher test
}
