% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald.R
\name{Wald}
\alias{Wald}
\title{To conduct Wald test}
\usage{
Wald(Z_score, Sigma)
}
\arguments{
\item{Z_score}{GWAS summary statistics}

\item{Sigma}{the correlation matrix}
}
\value{
pval.Wald
}
\description{
To conduct Wald test
}
